+++ CHANGELOG.md
@@ -9,17 +9,23 @@ Check [Changelog Format](#Changelog-Format) for more details about this changelo
 
 - Add details and URLs of the generated artifacts by the GitHub Actions
 - Append error details in the patch's comment in case of an error occurs
+- Parsing CI Broker's message protocol version
+- Support for CI Broker message protocol version 1
 
 ### Changed
 
 - Support single comment in a patch (instead of multiple) which is updated periodically with workflows' progress and 
   results
 
+### Removed
+
+- Broker's Error Response which supported the Error variant (without protocol version support)
+
 ### Fixed
 
 - Documentation issues and updates
 
-## [0.5.0] - 2024-03-21
+## [v0.5.0] - 2024-03-21
 
 ### Added